requirements.txt
# To activate this run
#
# pip install -r requirements.txt
# python -m django --version
#
# On a Macintosh this should be python3 and pip3
#
python==3.7.10
django==3.2.5
django-crispy-forms
django-filter
djangorestframework
Markdown
social-auth-app-django
social-auth-core
django-extensions
django-taggit
mysqlclient
django-ckeditor
Pillow
